destruction bug fixed? also on linux?

+++ src/Stopwatch.cpp
@@ -13,13 +13,13 @@ namespace utils {
 
     void Stopwatch::reset() {
         auto end = std::chrono::high_resolution_clock::now();
-        std::chrono::duration<double> time = end - start;
+        std::chrono::duration<float> time = end - start;
 
         frame_time = time.count();
         start = end;
     }
 
-    double Stopwatch::getFrame_time() const {
+    float Stopwatch::getFrame_time() const {
         return frame_time;
     }
 }
